requirements.txt
streamlit
Pillow
numpy
pandas
torch  # Required for YOLO and PyTorch model compatibility
ultralytics
opencv-python-headless>=4.5.5